backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.9.0
google-generativeai>=0.3.2
playwright>=1.41.0
httpx>=0.26.0
python-dotenv>=1.0.0
redis>=5.0.1
requests>=2.31.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0

